Unify logic for searching for instances
