+++ creusot/src/rustc_extensions.rs
@@ -1,4 +1,1 @@
 pub mod codegen;
-mod instance;
-
-pub use instance::*;
