+++ creusot/tests/should_succeed/drop_pair.stdout
@@ -42,12 +42,6 @@ module CreusotContracts_Builtins_Impl10_Resolve
   predicate resolve (self : (t1, t2)) = 
     Resolve0.resolve (let (a, _) = self in a) && Resolve1.resolve (let (_, a) = self in a)
 end
-module CreusotContracts_Builtins_Impl10
-  type t1   
-  type t2   
-  clone CreusotContracts_Builtins_Impl10_Resolve as Resolve0 with type t1 = t1, type t2 = t2
-  clone CreusotContracts_Builtins_Resolve with type self = (t1, t2), predicate resolve = Resolve0.resolve
-end
 module DropPair_DropPair_Interface
   use prelude.Prelude
   use mach.int.Int
@@ -118,12 +112,6 @@ module CreusotContracts_Builtins_Impl11_Resolve
   use prelude.Prelude
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
-end
-module CreusotContracts_Builtins_Impl11
-  type t   
-  use prelude.Prelude
-  clone CreusotContracts_Builtins_Impl11_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve with type self = borrowed t, predicate resolve = Resolve0.resolve
 end
 module DropPair_Drop_Interface
   use prelude.Prelude
