+++ creusot/tests/should_succeed/inc_max_3.stdout
@@ -25,12 +25,6 @@ module CreusotContracts_Builtins_Impl11_Resolve
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
 end
-module CreusotContracts_Builtins_Impl11
-  type t   
-  use prelude.Prelude
-  clone CreusotContracts_Builtins_Impl11_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve with type self = borrowed t, predicate resolve = Resolve0.resolve
-end
 module IncMax3_Swap_Interface
   use prelude.Prelude
   use mach.int.Int
