+++ creusot/tests/should_succeed/list_index_mut.stdout
@@ -76,12 +76,6 @@ module CreusotContracts_Builtins_Impl11_Resolve
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
 end
-module CreusotContracts_Builtins_Impl11
-  type t   
-  use prelude.Prelude
-  clone CreusotContracts_Builtins_Impl11_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve with type self = borrowed t, predicate resolve = Resolve0.resolve
-end
 module Std_Process_Abort_Interface
   val abort () : ()
     ensures { false }
@@ -313,11 +307,6 @@ module ListIndexMut_Write
   }
   
 end
-module ListIndexMut_Impl0
-  use Type
-  clone ListIndexMut_Impl0_Resolve as Resolve0
-  clone CreusotContracts_Builtins_Resolve with type self = Type.listindexmut_list, predicate resolve = Resolve0.resolve
-end
 module ListIndexMut_Main_Interface
   val main () : ()
 end
@@ -381,4 +370,9 @@ module ListIndexMut_Main
     return _0
   }
   
+end
+module ListIndexMut_Impl0
+  use Type
+  clone ListIndexMut_Impl0_Resolve as Resolve0
+  clone CreusotContracts_Builtins_Resolve with type self = Type.listindexmut_list, predicate resolve = Resolve0.resolve
 end
