+++ creusot/tests/should_succeed/modules.stdout
@@ -22,16 +22,6 @@ module Modules_Nested_Impl0_Resolve
   predicate resolve (self : Type.modules_nested_nested) = 
     true
 end
-module CreusotContracts_Builtins_Resolve
-  type self   
-  predicate resolve (self : self)
-end
-module Modules_Nested_Impl0
-  use Type
-  clone Modules_Nested_Impl0_Resolve as Resolve0
-  clone CreusotContracts_Builtins_Resolve with type self = Type.modules_nested_nested,
-  predicate resolve = Resolve0.resolve
-end
 module Modules_Nested_InnerFunc_Interface
   val inner_func () : bool
     ensures { result = true }
@@ -98,4 +88,14 @@ module Modules_Main
     return _0
   }
   
+end
+module CreusotContracts_Builtins_Resolve
+  type self   
+  predicate resolve (self : self)
+end
+module Modules_Nested_Impl0
+  use Type
+  clone Modules_Nested_Impl0_Resolve as Resolve0
+  clone CreusotContracts_Builtins_Resolve with type self = Type.modules_nested_nested,
+  predicate resolve = Resolve0.resolve
 end
