+++ creusot/tests/should_succeed/while_let.stdout
@@ -29,12 +29,6 @@ module CreusotContracts_Builtins_Impl11_Resolve
   predicate resolve (self : borrowed t) = 
      ^ self =  * self
 end
-module CreusotContracts_Builtins_Impl11
-  type t   
-  use prelude.Prelude
-  clone CreusotContracts_Builtins_Impl11_Resolve as Resolve0 with type t = t
-  clone CreusotContracts_Builtins_Resolve with type self = borrowed t, predicate resolve = Resolve0.resolve
-end
 module WhileLet_Main_Interface
   val main () : ()
 end
